requirements.txt
albumentations==1.3.1
cv2
joblib==1.3.2
matplotlib==3.8.2
natsort==8.4.0
nibabel==5.2.1
numpy==1.24.4
pandas==1.5.3
scipy==1.11.3
skimage
sklearn
streamlit==1.36.0
torch==2.3.0
torchmetrics==1.3.0
torchvision==0.16.0
tqdm==4.66.2
pyyaml
